Note insecurity of login storage
Warn users that, if they use the "remember me" function, it's not
secure, per Alys/tools-for-habitrpg#24.  Possibly this text is overly
strong, but I'd rather it be overly strong than insufficiently cautious,
and given the feature is very unlikely to ever be merged in anyway, it
doesn't matter if it's too offputting for other users.